requirements.txt
discord.py @ https://github.com/Rapptz/discord.py
pillow
opencv-python
opencv-contrib-python
python-dotenv
